+++ samples/Demo.HelloAvalonia/MainWindowViewModel.cs
@@ -1,0 +1,8 @@
+﻿using Caliburn.Light;
+
+namespace Demo.HelloAvalonia;
+
+public sealed partial class MainWindowViewModel : Screen
+{
+    public string Greeting { get; } = "Welcome to Avalonia!";
+}