+++ src/Caliburn.Light.Avalonia/ExecutionContextResolver.cs
@@ -1,0 +1,38 @@
+using Avalonia;
+
+namespace Caliburn.Light.Avalonia;
+
+/// <summary>
+/// Special value to resolve <see cref="CommandExecutionContext"/> parameter.
+/// </summary>
+public sealed class ExecutionContextResolver : AvaloniaObject, ISpecialValue
+{
+    /// <summary>
+    /// Identifies the <seealso cref="ExecutionContextResolver.SourceOverride"/> dependency property.
+    /// </summary>
+    public static readonly StyledProperty<object?> SourceOverrideProperty = AvaloniaProperty.Register<ExecutionContextResolver, object?>(
+        "SourceOverride");
+
+    /// <summary>
+    /// Gets or sets the override for <see cref="CommandExecutionContext.Source"/>.
+    /// </summary>
+    public object? SourceOverride
+    {
+        get { return GetValue(SourceOverrideProperty); }
+        set { SetValue(SourceOverrideProperty, value); }
+    }
+
+    /// <summary>
+    /// Resolves the value of this instance.
+    /// </summary>
+    /// <param name="context">The context.</param>
+    /// <returns>The resolved value.</returns>
+    public object? Resolve(CommandExecutionContext context)
+    {
+        var sourceOverride = SourceOverride;
+        if (sourceOverride is not null)
+            context.Source = sourceOverride;
+
+        return context;
+    }
+}