+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- `IWindowManager.ShowDialog(viewModel, ownerViewModel, context)` throws when owner Window cannot be determined
+
 ### Removed
 - WPF: Removed `IWindowManager.ShowPopup(viewModel, context)`